Use indirect coordinate system using pointers

+++ imcore/src/main/java/es/ua/dlsi/im3/core/score/layout/svg/SVGExporter.java
@@ -42,8 +42,8 @@ public class SVGExporter implements IGraphicsExporter {
         XMLExporterHelper.start(sb, 0, "svg",
                 "version", "1.1",
                 "baseProfile", "full",
-                "width", Integer.toString(canvas.getWidth()),
-                "height", Integer.toString(canvas.getHeight()),
+                "width", Integer.toString((int) canvas.getWidth()),
+                "height", Integer.toString((int) canvas.getHeight()),
                 "xmlns", "http://www.w3.org/2000/svg",
                 "xmlns:xlink", "http://www.w3.org/1999/xlink",
                 "xml:space", "preserve"
